+++ src/data/v2/quotes.rs
@@ -36,7 +36,7 @@ pub struct Quotes {
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct QuotesReqInit {
   /// See `QuotesReq::limit`.
-  pub limit: Option<u64>,
+  pub limit: Option<usize>,
   /// See `QuotesReq::page_token`.
   pub page_token: Option<String>,
   #[doc(hidden)]
@@ -78,7 +78,7 @@ pub struct QuotesReq {
   /// Number of quotes to return. Must be in range 1-10000, defaults to
   /// 1000.
   #[serde(rename = "limit")]
-  pub limit: Option<u64>,
+  pub limit: Option<usize>,
   /// Pagination token to continue from.
   #[serde(rename = "page_token")]
   pub page_token: Option<String>,
@@ -112,7 +112,7 @@ Endpoint! {
     struct Query<'s> {
       start: DateTime<Utc>,
       end: DateTime<Utc>,
-      limit: Option<u64>,
+      limit: Option<usize>,
       page_token: Option<&'s str>,
     }
